Commit for trailofbits aduit

+++ packages/loopring_v3/contracts/aux/agents/FastWithdrawalAgent.sol
@@ -53,7 +53,7 @@ contract FastWithdrawalAgent is ReentrancyGuard, IAgent
         address from,
         address to,
         address token,
-        uint96  amount,
+        uint248  amount,
         address provider,
         bool    success
     );
@@ -64,7 +64,7 @@ contract FastWithdrawalAgent is ReentrancyGuard, IAgent
         address from;                   // The owner of the account
         address to;                     // The `to` address of the withdrawal
         address token;
-        uint96  amount;
+        uint248  amount;
         uint32  storageID;
     }
